Remove 'load' method from 'Harness'.
